fix: rename grid options  to grid settings
