+++ apps/admin/src/plugins/pageBuilder/renderPlugins.ts
@@ -26,7 +26,7 @@ import align from "@webiny/app-page-builder/render/plugins/elementSettings/align
 import animation from "@webiny/app-page-builder/render/plugins/elementSettings/animation";
 import background from "@webiny/app-page-builder/render/plugins/elementSettings/background";
 import border from "@webiny/app-page-builder/render/plugins/elementSettings/border";
-import gridOptions from "@webiny/app-page-builder/render/plugins/elementSettings/grid";
+import gridSettings from "@webiny/app-page-builder/render/plugins/elementSettings/grid";
 import height from "@webiny/app-page-builder/render/plugins/elementSettings/height";
 import width from "@webiny/app-page-builder/render/plugins/elementSettings/width";
 import shadow from "@webiny/app-page-builder/render/plugins/elementSettings/shadow";
@@ -66,7 +66,7 @@ export default [
     animation,
     background,
     border,
-    gridOptions,
+    gridSettings,
     height,
     width,
     shadow,
